+++ Surface_reconstruction_3/test/Surface_reconstruction_3/outliers_removal_test.cpp
@@ -46,7 +46,7 @@ typedef Kernel::Point_3 Point;
 
 void test_avg_knn_sq_distance(std::deque<Point>& points, // input point set
                               double nb_neighbors_outliers_removal, // number of neighbors
-                              double threshold_percent_avg_knn_sq_dst) // percentage of points to remove 
+                              double threshold_percent_avg_knn_sq_dst) // percentage of points to remove
 {
   CGAL::Timer task_timer; task_timer.start();
 
@@ -54,24 +54,24 @@ void test_avg_knn_sq_distance(std::deque<Point>& points, // input point set
   int nb_neighbors = int(double(points.size()) * nb_neighbors_outliers_removal / 100.0);
   if (nb_neighbors < 7)
     nb_neighbors = 7;
-  if (nb_neighbors > points.size()-1)
+  if ((unsigned int)nb_neighbors > points.size()-1)
     nb_neighbors = points.size()-1;
 
   std::cerr << "Remove outliers wrt average squared distance to knn (remove "
-            << threshold_percent_avg_knn_sq_dst << "%, knn=" 
+            << threshold_percent_avg_knn_sq_dst << "%, knn="
             << nb_neighbors_outliers_removal << "%=" << nb_neighbors << ")...\n";
 
   std::deque<Point> output;
   CGAL::remove_outliers_wrt_avg_knn_sq_distance_3(
                   points.begin(), points.end(),
                   std::back_inserter(output),
-                  nb_neighbors, 
+                  nb_neighbors,
                   threshold_percent_avg_knn_sq_dst);
-  
+
   // mutating version of the same function
   points.erase(CGAL::remove_outliers_wrt_avg_knn_sq_distance_3(
                   points.begin(), points.end(),
-                  nb_neighbors, 
+                  nb_neighbors,
                   threshold_percent_avg_knn_sq_dst),
                points.end());
 
@@ -114,12 +114,12 @@ int main(int argc, char * argv[])
     // Load point set
     std::deque<Point> points;
     std::cerr << "Open " << argv[i] << " for reading...";
-    if(CGAL::surface_reconstruction_read_xyz(argv[i], 
-                                             std::back_inserter(points), 
+    if(CGAL::surface_reconstruction_read_xyz(argv[i],
+                                             std::back_inserter(points),
                                              false /*skip normals*/))
     {
       std::cerr << "ok (" << points.size() << " points)" << std::endl;
-      
+
       test_avg_knn_sq_distance(points, nb_neighbors_outliers_removal, threshold_percent_avg_knn_sq_dst);
     }
     else
@@ -135,4 +135,4 @@ int main(int argc, char * argv[])
   std::cerr << "Tool returned " << accumulated_fatal_err << std::endl;
   return accumulated_fatal_err;
 }
- 
+
